fixing warning #3

+++ src/utils/iodevices/OutputDevice.cpp
@@ -232,9 +232,10 @@ OutputDevice::setPrecision(int precision) {
     getOStream() << std::setprecision(precision);
 }
 
+
 int
 OutputDevice::precision() {
-    return getOStream().precision();
+    return (int)getOStream().precision();
 }
